exyract functionality of edit for a single recipe into a new component
